New TRRReader class
Created a reader for trr files in analogy to the XTCReader.
Frames are returned as TRRFrames, which have additonal fields
'velocities' and 'forces', which can be stored in trr files.

Creation of index files is slightly different to xtcfiles:
When the TRRReader class is instantiated with the indexfile argument,
the cache will be loaded from this file, or created if the file does not
exist. This is handled by the function pygmx.open().